+++ query-graphs-js/src/graphPath.ts
@@ -206,7 +206,7 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
       edgeConditions: [],
       ownPathIds: [],
       overriddingPathIds: [],
-      runtimeTypesOfTail: runtimeTypes
+      runtimeTypesOfTail: runtimeTypes,
     });
   }
 
@@ -281,6 +281,18 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
   /** The possible runtime types the tail of the path can be (this is deduplicated). */
   tailPossibleRuntimeTypes(): readonly ObjectType[] {
     return this.props.runtimeTypesOfTail;
+  }
+
+  /**
+   * Returns `true` if the last edge of the path correspond to an @interfaceObject "fake cast" while the the previous edge was an edge that "entered" the subgraph (a key edge from another subgraph).
+   */
+  lastIsIntefaceObjectFakeDownCastAfterEnteringSubgraph(): boolean {
+    return this.lastIsInterfaceObjectFakeDownCast()
+      && this.subgraphEnteringEdge?.index === this.size - 2; // size - 1 is the last index (the fake cast), so size - 2 is the previous edge.
+  }
+
+  private lastIsInterfaceObjectFakeDownCast(): boolean {
+    return this.lastEdge()?.transition.kind === 'InterfaceObjectFakeDownCast';
   }
 
   /**
@@ -368,13 +380,14 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
         // type is not an @interfaceObject itself, then we can eliminate that last edge as it does nothing useful, but also,
         // it has conditions and we don't need/want the key we're following to depend on those conditions, since it doesn't have
         // to.
-        if (this.props.edgeToTail?.transition.kind === 'InterfaceObjectFakeDownCast' && isInterfaceType(edge.tail.type)) {
+        if (this.lastIsInterfaceObjectFakeDownCast() && isInterfaceType(edge.tail.type)) {
           return new GraphPath({
             ...this.props,
             tail: edge.tail,
             edgeTriggers: withReplacedLastElement(this.props.edgeTriggers, trigger),
             edgeIndexes: withReplacedLastElement(this.props.edgeIndexes, edge.index),
             edgeConditions: withReplacedLastElement(this.props.edgeConditions, conditionsResolution.pathTree ?? null),
+            subgraphEnteringEdge,
             edgeToTail: edge,
             runtimeTypesOfTail: updateRuntimeTypes(this.props.runtimeTypesOfTail, edge),
             runtimeTypesBeforeTailIfLastIsCast: undefined, // we know last is not a cast
@@ -436,7 +449,7 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
     });
   }
 
-  checkDirectPathFomPreviousSubgraphTo(
+  checkDirectPathFromPreviousSubgraphTo(
     typeName: string,
     triggerToEdge: (graph: QueryGraph, vertex: Vertex, t: TTrigger) => Edge | null | undefined
   ): Vertex | undefined {
@@ -1116,6 +1129,25 @@ function advancePathWithNonCollectingAndTypePreservingTransitions<TTrigger, V ex
   convertTransitionWithCondition: (transition: Transition, context: PathContext) => TTrigger,
   triggerToEdge: (graph: QueryGraph, vertex: Vertex, t: TTrigger) => Edge | null | undefined
 ): IndirectPaths<TTrigger, V, TNullEdge, TDeadEnds>  {
+  // If we're asked for indirect paths after an "@interfaceObject fake down cast" but that down cast comes just after a non-collecting edges, then
+  // we can ignore it (skip indirect paths from there). The reason is that the presence of the non-collecting just before the fake down-cast means
+  // we add looked at indirect paths just before that down cast, but that fake downcast really does nothing in practice with the subgraph it's on,
+  // so any indirect path from that fake down cast will have a valid indirect path _before_ it, and so will have been taken into account independently.
+  if (path.lastIsIntefaceObjectFakeDownCastAfterEnteringSubgraph()) {
+    // Note: we need to register a dead-end for every subgraphs we "could" be going to, or the code calling this may try to infer a reason on its own
+    // and we'll run into some assertion.
+    const reachableSubgraphs = new Set(path.nextEdges().filter((e) => !e.transition.collectOperationElements && e.tail.source !== path.tail.source).map((e) => e.tail.source));
+    return {
+      paths: [],
+      deadEnds: new Unadvanceables(Array.from(reachableSubgraphs).map((s) => ({
+        sourceSubgraph: path.tail.source,
+        destSubgraph: s,
+        reason: UnadvanceableReason.IGNORED_INDIRECT_PATH,
+        details: `ignoring moving from "${path.tail.source}" to "${s}" as a more direct option exists`,
+      }))) as TDeadEnds,
+    };
+  }
+
   const isTopLevelPath = path.isOnTopLevelQueryRoot();
   const typeName = isFederatedGraphRootType(path.tail.type) ? undefined : path.tail.type.name;
   const originalSource = path.tail.source;
@@ -1251,7 +1283,7 @@ function advancePathWithNonCollectingAndTypePreservingTransitions<TTrigger, V ex
         // loop when calling `hasValidDirectKeyEdge` in that case without additional care and it's not useful because this
         // very method already ensure we don't create unnecessary chains of keys for the "current type"
         if (subgraphEnteringEdge && edge.transition.kind === 'KeyResolution' && subgraphEnteringEdge.edge.tail.type.name !== typeName) {
-          const prevSubgraphVertex = toAdvance.checkDirectPathFomPreviousSubgraphTo(edge.tail.type.name, triggerToEdge);
+          const prevSubgraphVertex = toAdvance.checkDirectPathFromPreviousSubgraphTo(edge.tail.type.name, triggerToEdge);
           const backToPreviousSubgraph = subgraphEnteringEdge.edge.head.source === edge.tail.source;
           const maxCost = toAdvance.subgraphEnteringEdge.cost + (backToPreviousSubgraph ? 0 : conditionResolution.cost);
           if (prevSubgraphVertex