+++ gateway-js/CHANGELOG.md
@@ -18,6 +18,8 @@ This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The
   intersection at all [PR #1556](https://github.com/apollographql/federation/pull/1556). 
 - Adds support for the 0.3 version of the tag spec, which adds `@tag` directive support for the `SCHEMA` location [PR #2314](https://github.com/apollographql/federation/pull/2314).
 - Fix potential issue with nested `@defer` in non-deferrable case [PR #2312](https://github.com/apollographql/federation/pull/2312).
+- Fixes composition issues with `@interfaceObject` [PR #2318](https://github.com/apollographql/federation/pull/2318).
+- Improves generation of plans once all path options are computed [PR #2316](https://github.com/apollographql/federation/pull/2316).
 
 ## 2.2.2